backend/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.3
langchain-community>=0.0.10
fastapi>=0.68.0
python-dotenv>=0.19.0
uvicorn>=0.15.0
python-multipart>=0.0.5
reportlab>=3.6.8
